pyproject.toml
[project]
name = "pense-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.127.0",
    "uvicorn>=0.40.0",
    "boto3>=1.34.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "openpyxl>=3.1.5",
]
